% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascvd_acc_aha.R
\name{ascvd_acc_aha}
\alias{ascvd_acc_aha}
\title{ACC/AHA 2013 ASCVD risk score for people aged 40-79 years}
\usage{
ascvd_acc_aha(ethnicity = c("white", "aa"), sex, age,
totchol, hdl, sbp, bp_med, smoker, diabetic)
}
\arguments{
\item{ethnicity}{ethnicity; categorical [white|aa]; ("aa"=afro american)}

\item{sex}{gender; categorical [female|male]}

\item{age}{age; integer [years]}

\item{totchol}{total cholesterol; numeric [mg/dl]}

\item{hdl}{high-density lipoprotein; numeric [mg/dl]}

\item{sbp}{systolic blood pressure; numeric [mmHg]}

\item{bp_med}{information if individual is on a blood pressure medication; numeric [1|0]; ("1"=yes;"0"=no)}

\item{smoker}{information on current self-reported smoking status; numeric [1|0]; ("1"=smoker;"0"=non-smoker)}

\item{diabetic}{diabetic status of individual; numeric [1|0]; ("1"=diabetic;"0"=non-diabetic)}
}
\value{
Estimated 10 year risk for hard ASCVD (percent)
}
\description{
Computes 10-year risk for hard atherosclerotic cardiovascular disease (ASCVD) event (defined as first occurrence of
non-fatal myocardial infarction (MI), congestive heart disease (CHD) death,
or fatal or nonfatal stroke).
}
\details{
The ASCVD (atherosclerotic cardiovascular disease) risk score is a national guideline developed by the American College of Cardiology.
It is a calculation of your 10-year risk of having a cardiovascular problem, such as a heart attack or stroke.
This risk estimate considers age, sex, ethnicity, cholesterol levels, blood pressure, medication use, diabetic status, and smoking status.
}
\references{
Goff, David C., et al. "2013 ACC/AHA guideline on the assessment of
cardiovascular risk: a report of the American College of
Cardiology/American Heart Association Task Force on Practice
Guidelines." Journal of the American College of Cardiology 63.25
Part B (2014): 2935-2959.
}
