% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascvd_frs_chd.R
\name{ascvd_frs_chd_formula}
\alias{ascvd_frs_chd_formula}
\alias{ascvd_frs_chd_table}
\title{Framingham risk score (CHD)}
\usage{
ascvd_frs_chd_formula(sex, age, totchol = NA, hdl, ldl = NA, sbp, dbp, smoker, diabetic, chol_cat = c("tc", "ldl"))
ascvd_frs_chd_table(sex, age, totchol = NA, hdl, ldl = NA, sbp, dbp, smoker, diabetic, chol_cat = c("tc", "ldl"))
}
\arguments{
\item{sex}{gender; categorical [female|male]}

\item{age}{age; integer [years]}

\item{totchol}{total cholesterol; numeric [mg/dl]}

\item{hdl}{high-density lipoprotein; numeric [mg/dl]}

\item{ldl}{low-density lipoprotein; numeric [mg/dl]}

\item{sbp}{systolic blood pressure; numeric [mmHg]}

\item{dbp}{diastolic blood pressure; numeric [mmHg]}

\item{smoker}{information on current self-reported smoking status; numeric [1|0]; ("1"=smoker;"0"=non-smoker)}

\item{diabetic}{diabetic status of individual; numeric [1|0]; ("1"=diabetic;"0"=non-diabetic)}

\item{chol_cat}{value to specify the use of total cholesterol ("tc") or LDL cholesterol ("ldl") in the function; categorical [tc|ldl]}
}
\value{
Estimated 10 year risk for CHD (percent)
}
\description{
Computes 10-year risk of developing coronary heart disease (CHD).
}
\details{
Background: \cr
The objective of this study was to examine the association of Joint National Committee (JNC-V) blood pressure and National Cholesterol Education Program (NCEP) cholesterol categories
with coronary heart disease (CHD) risk, to incorporate them into coronary prediction algorithms, and to compare the discrimination properties of this approach
with other noncategorical prediction functions.\cr
Methods and Results:\cr
This work was designed as a prospective, single-center study in the setting of a community-based cohort. The patients were 2489 men and 2856 women 30 to 74 years old at baseline
with 12 years of follow-up. During the 12 years of follow-up, a total of 383 men and 227 women developed CHD, which was significantly associated with categories of
blood pressure, total cholesterol, LDL cholesterol, and HDL cholesterol (all p < 0.001). Sex-specific prediction equations were formulated to predict CHD risk according to
age, diabetes, smoking, JNC-V blood pressure categories, and NCEP total cholesterol and LDL cholesterol categories. The accuracy of this categorical approach was found
to be comparable to CHD prediction when the continuous variables themselves were used. After adjustment for other factors, 28\% of CHD events in men and 29\% in women
were attributable to blood pressure levels that exceeded high normal (130/85). The corresponding multivariable-adjusted attributable risk percent associated with
elevated total cholesterol (200 mg/dL) was 27\% in men and 34\% in women.\cr
Conclusions:\cr
Recommended guidelines of blood pressure, total cholesterol, and LDL cholesterol effectively predict CHD risk in a middle-aged white population sample.
A simple coronary disease prediction algorithm was developed using categorical variables, which allows physicians to predict multivariate CHD risk in patients without overt CHD.
}
\references{
Wilson PW, et al.Prediction of coronary heart disease using risk factor categories. Circulation. 1998 May 12;97(18):1837-47. doi: 10.1161/01.cir.97.18.1837. PMID: 9603539.
}
