% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESC_Score2016_table.R
\name{ESC_SCORE2016table}
\alias{ESC_SCORE2016table}
\title{Calculate ESC-Score 2016 Table Version}
\usage{
ESC_SCORE2016table(cholesterol, gender, age, RRsys, smoking)
}
\arguments{
\item{cholesterol}{Cholesterin REDCap-ID: varid_1206; in mg/dl}

\item{gender}{Geschlecht REDCap-ID: varid_549}

\item{age}{Alter REDCap-ID: varid_1891}

\item{RRsys}{Blutdruck nach 5 min Ruhe - systolisch REDCap-ID: varid_627; in mmHg}

\item{smoker}{Raucher REDCap-ID: varid_561}
}
\value{
A vector of the calculated risk per record.
}
\description{
This function takes necessary parameters to calculate the ESC-Score
}
\details{
Die SCORE (Systematic Coronary Risk Evaluation) Risikodiagramme werden für die absoluten 10-Jahres Risikoschätzungen für eine
tödliche Herz-Kreislauf-Erkrankung einschließlich Schlaganfällen eingesetzt bei Patienten, die eine Primärprävention benötigen (Rücker et al., 2016).
Die Risikofaktoren die der Score zur Berechnung nimmt sind: Alter, Geschlecht, systolischer Blutdruck (SBP) in mmHg,
Cholesterin (Gesamtcholesterin oder das HDL-C Verhältnis) in der Einheit mg/dl und Rauchen.
Im Originalscore wird das Alter eingegrenzt auf 40-65 Jahre und in 5-Jahresabschnitten eingestuft, in dieser Funktion gibt es keine Eingrenzung des Alters.
Um das absolute 10-Jahres-Risiko eines Herz-Kreislauf-Todes für eine Person abzuschätzen,
muss ihr Alter auf das in der Tabelle angegebene Alter und die dem systolischen Blutdruck und Gesamtcholesterin der Person am nächsten liegende Raucher- oder Nichtraucherzelle
gerundet werden. Die Einteilung kann man der Funktion entnehmen.
}
