% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESC_Score_2016_table.R
\name{ESC_Score_2016_table}
\alias{ESC_Score_2016_table}
\title{Calculate ESC-Score 2016 table version}
\usage{
ESC_Score_2016_table(sex, age, totchol, sbp, smoker, risk = c("low","high"), mmol = FALSE)
}
\arguments{
\item{sex}{gender; categorical [female|male]}

\item{age}{age; integer [years]}

\item{totchol}{total cholesterol; numeric [mg/dl] or [mmol/L]. If unit is mg/dL set the argument mmol=FALSE}

\item{sbp}{systolic blood pressure; numeric [mmHg]}

\item{smoker}{information on current self-reported smoking status; numeric [1|0]; ("1"=smoker;"0"=non-smoker)}

\item{risk}{choose which risk chart is used for calculation; categorical ["low"|"high"]}

\item{mmol}{is Cholesterol given as mmol/L; logical [TRUE|FALSE]}
}
\value{
A vector of calculated risks of persons.
}
\description{
This function takes necessary parameters to calculate the ESC-Score 2016 table version for people aged 40-65 years in two geographical risk regions.
}
\details{
The SCORE risk assessment is derived from a large dataset of prospective European studies and predicts fatal atherosclerotic CVD events over a ten year period.
This risk estimation is based on the following risk factors: sex, age, smoker, systolic blood pressure and total cholesterol.
The threshold for high risk based on fatal cardiovascular events is defined as "higher than 5\%", instead of the previous "higher than 20\%" using a composite coronary endpoint.
This SCORE model has been calibrated according to each European country’s mortality statistics. In other words, if used on the entire population aged 40-65, it will predict the exact number of fatal CVD-events that eventually will occur after 10 years.
The relative risk chart may be used to show younger people at low total risk that, relative to others in their age group, their risk may be many times higher than necessary. This may help to motivate decisions about avoidance of smoking, healthy nutrition and exercise, as well as flagging those who may become candidates for medication. This chart refers to relative risk, not percentage risk.
You can read more about the SCORE project in European Heart Journal, 2003, 24; 987-1003.
}
\references{
Piepoli MF,et al. 2016 European Guidelines on cardiovascular disease prevention in clinical practice: The Sixth Joint Task Force of the European Society of Cardiology and Other Societies on Cardiovascular Disease Prevention in Clinical Practice (constituted by representatives of 10 societies and by invited experts):
Developed with the special contribution of the European Association for Cardiovascular Prevention & Rehabilitation (EACPR).
Eur J Prev Cardiol. 2016 Jul;23(11):NP1-NP96. doi: 10.1177/2047487316653709. Epub 2016 Jun 27. PMID: 27353126.
}
