% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESC_Score_GER_2016_table.R
\name{ESC_Score_GER_2016_table}
\alias{ESC_Score_GER_2016_table}
\title{Calculate ESC-Score Germany 2016 Table Version}
\usage{
ESC_Score_GER_2016_table(totchol, sex, age, sbp, smoker, mmol = FALSE)
}
\arguments{
\item{sex}{a character vector indicating the sex of the person. Values: "female", "male"}

\item{age}{a numeric vector with the age of persons given in years}

\item{totchol}{a numeric vector; Cholesterol values given in mg/dL or mmol/L. If unit is mg/dL set  the argument mmol to FALSE}

\item{sbp}{a numeric vector with the systolic blood pressure of persons given as mmHg}

\item{smoker}{a numeric vector. Smoker = 1, non-smoker = 0. A smoker was defined as current self-reported smoker.}

\item{mmol}{logical. Is Cholesterol given as mmol/L (TRUE) or mg/dL (FALSE).}
}
\value{
A vector of the calculated risk per record.
}
\description{
This function takes necessary parameters to calculate the ESC-Score Germany 2016 Table Version
}
\details{
Abstract: Estimation of absolute risk of cardiovascular disease (CVD), preferably with population-specific risk charts, has become a cornerstone of CVD primary prevention.
Regular recalibration of risk charts may be necessary due to decreasing CVD rates and CVD risk factor levels. The SCORE risk charts for fatal CVD risk assessment were first calibrated
for Ger- many with 1998 risk factor level data and 1999 mortality statistics. We present an update of these risk charts based on the SCORE methodology including estimates of relative
risks from SCORE, risk factor levels from the German Health Interview and Examination Survey for Adults 2008–11 (DEGS1) and official mortality statistics from 2012. Competing risks methods
were applied and estimates were independently validated. Updated risk charts were calculated based on cholesterol, smoking, systolic blood pressure risk factor levels, sex and 5-year age-groups.
The absolute 10-year risk estimates of fatal CVD were lower according to the updated risk charts compared to the first calibration for Germany. In a nationwide sample of 3062 adults
aged 40–65 years free of major CVD from DEGS1, the mean 10-year risk of fatal CVD estimated by the updated charts was lower by 29\% and the estimated proportion of high
risk people (10-year risk > = 5\%) by 50\% compared to the older risk charts. This recalibration shows a need for regular updates of risk charts according to changes in mortality and
risk factor levels in order to sustain the identification of people with a high CVD risk.
}
\references{
Rücker V, Keil U, Fitzgerald AP, Malzahn U, Prugger C, Ertl G, Heuschmann PU, Neuhauser H.
Predicting 10-Year Risk of Fatal Cardiovascular Disease in Germany: An Update Based on the SCORE-Deutschland Risk Charts.
PLoS One. 2016 Sep 9;11(9):e0162188. doi: 10.1371/journal.pone.0162188. PMID: 27612145; PMCID: PMC5017762.
}
