% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESC_Score_OP_table.R
\name{ESC_Score_OP_table}
\alias{ESC_Score_OP_table}
\title{Calculate ESC-Score OP Table Version}
\usage{
ESC_Score_OP_table(totchol, sex, age, sbp, smoker, risk = c("low","high"), mmol = FALSE)
}
\arguments{
\item{totchol}{a numeric vector; Cholesterol values given in mg/dL or mmol/L. If unit is mg/dL set  the argument mmol to FALSE}

\item{sex}{a numeric vector indicating the sex of the person. Values: "female" = 1, "male" = 0}

\item{age}{a numeric vector with the age of persons given as years}

\item{sbp}{a numeric vector with the systolic blood pressure of persons given as mmHg}

\item{smoker}{a numeric vector. Smoker = 1, non-smoker = 0. A smoker was defined as current self-reported smoker.}

\item{risk}{logical. Choose if which risk chart is used for calculation}

\item{mmol}{logical. Is Cholesterol given as mmol/L (TRUE) or mg/dL (FALSE).}
}
\value{
A vector of the calculated risk per record.
}
\description{
This function takes necessary parameters to calculate the ESC-Score Older People (OP) Table Version for high and low risk
}
\details{
Abstract: Aims Estimation of cardiovascular disease risk, using SCORE (Systematic Coronary Risk Evaluation) is recommended by European guidelines on cardiovascular disease prevention.
Risk estimation is inaccurate in older people. We hypothesized that this may be due to the assumption, inherent in current risk estimation systems, that risk factors function similarly
in all age groups. We aimed to derive and validate a risk estimation function, SCORE O.P., solely from data from individuals aged 65 years and older.
Methods and results 20,704 men and 20,121 women, aged 65 and over and without pre-existing coronary disease, from four representative, prospective studies of the general population were included.
These were Italian, Belgian and Danish studies (from original SCORE dataset) and the CONOR (Cohort of Norway) study. The variables which remained statistically significant
in Cox proportional hazards model and were included in the SCORE O.P. model were: age, total cholesterol, high-density lipoprotein cholesterol, systolic blood pressure, smoking status and diabetes.
SCORE O.P. showed good discrimination; area under receiver operator characteristic curve (AUROC) 0.74 (95\% confidence interval: 0.73 to 0.75). Calibration was also reasonable,
Hosmer-Lemeshow goodness of fit test: 17.16 (men), 22.70 (women). Compared with the original SCORE function extrapolated to the ≥65 years age group discrimination improved,
p = 0.05 (men), p < 0.001 (women). Simple risk charts were constructed. On simulated external validation, performed using 10-fold cross validation, AUROC was 0.74 and predicted/observed ratio
was 1.02. Conclusion SCORE O.P. provides improved accuracy in risk estimation in older people and may reduce excessive use of medication in this vulnerable population.
}
\references{
Cooney MT, et al. Cardiovascular risk estimation in older persons: SCORE O.P. Eur J Prev Cardiol. 2016 Jul;23(10):1093-103. doi: 10.1177/2047487315588390. Epub 2015 Jun 3. PMID: 26040999.
}
