% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESCScore_OP.R
\name{ESC_ScoreOPtable}
\alias{ESC_ScoreOPtable}
\title{Calculate ESC-Score OP Table Version}
\usage{
ESC_SCORE2016table(cholesterol, gender, age, RRsys, smoking)
}
\arguments{
\item{cholesterol}{Cholesterin REDCap-ID: varid_1206; in mg/dl}

\item{gender}{Geschlecht REDCap-ID: varid_549}

\item{age}{Alter REDCap-ID: varid_1891}

\item{RRsys}{Blutdruck nach 5 min Ruhe - systolisch REDCap-ID: varid_627; in mmHg}

\item{smoking}{Raucher REDCap-ID: varid_561}
}
\value{
A vector of the calculated risk per record.
}
\description{
This function takes necessary parameters to calculate the ESC-Score Older People (OP) Table Version for high and low risk
}
\details{
The SCORE risk assessment is derived from a large dataset of prospective European studies and predicts fatal atherosclerotic CVD events over a ten year period.
This risk estimation is based on the following risk factors: gender, age, smoking, systolic blood pressure and total cholesterol.
The threshold for high risk based on fatal cardiovascular events is defined as "higher than 5\%", instead of the previous "higher than 20\%" using a composite coronary endpoint.
This SCORE model has been calibrated according to each European country’s mortality statistics. In other words, if used on the entire population aged 40-65, it will predict the exact number of fatal CVD-events that eventually will occur after 10 years.
The relative risk chart may be used to show younger people at low total risk that, relative to others in their age group, their risk may be many times higher than necessary. This may help to motivate decisions about avoidance of smoking, healthy nutrition and exercise, as well as flagging those who may become candidates for medication. This chart refers to relative risk, not percentage risk.
You can read more about the SCORE project in European Heart Journal, 2003, 24; 987-1003.
}
