% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invest_score.R
\name{invest_score}
\alias{invest_score}
\title{Calculate INVEST-Score}
\usage{
invest_score(age = NA, ethnicity = NA, bmi = NA, hr = NA, sbp = NA, mi = NA, chf = NA, stroke = NA, smoker = NA, diabetic = NA, pad = NA, ckd = NA)
}
\arguments{
\item{age}{age; integer [years]}

\item{ethnicity}{ethnicity; categorical [white|nw]; ("nw"= nonwhite)}

\item{bmi}{body mass index; numeric [kg/m^2]}

\item{hr}{heart rate; numeric [beats/minute]}

\item{sbp}{systolic blood pressure; numeric [mmHg]}

\item{mi}{information of individual had a prior myocardial infarction (MI); numeric [1|0]; ("1"=yes;"0"=no)}

\item{chf}{information if individual has a congestive heart failure (CHF); numeric [1|0]; ("1"=yes;"0"=no)}

\item{stroke}{information if individual had a stroke; numeric [1|0]; ("1"=yes;"0"=no)}

\item{smoker}{a numeric vector. Smoker = 1, non-smoker = 0. A smoker was defined as current self-reported smoker.}

\item{diabetic}{diabetic status of individual; numeric [1|0]; ("1"=diabetic;"0"=non-diabetic)}

\item{pad}{information if individual has a peripheral arterial disease (PAD); numeric [1|0]; ("1"=yes;"0"=no)}

\item{ckd}{information if individual has a chronic kidney disease (CKD); numeric [1|0]; ("1"=yes;"0"=no)}
}
\value{
A vector of the calculated risk per record.
}
\description{
This function takes clinical parameters to calculate the INVEST-Score. A risk score for predicting important adverse events in patients with hypertension and chronic stable coronary artery disease (CAD)
}
\details{
The function takes 12 clinical variables to calculate a risk score (\%) to predict future adverse Events (Mortality all-cause, Non-fatal myocardial infarction, non-fatal stroke)
Patients can be divided into the following categories based on their individual risk score:
\itemize{
\item{Low risk}{Risk Score <= 4}
\item{Intermediate risk}{Risk Score 5 to 6}
\item{High risk}{Risk Score >= 12}
}
BACKGROUND:\cr
It is difficult to accurately determine prognosis of patients with hypertension and chronic stable coronary artery disease (CAD).
Our aim was to construct a risk score for predicting important adverse events in this population.\cr
METHODS and Results:\cr
Patients with hypertension and chronic stable CAD enrolled in the INternational VErapamil‐SR/Trandolapril STudy (INVEST) comprised the study cohort.
Candidate predictor variables were obtained from patients with at least 1 postbaseline visit.
Patients were divided into development (n=18 484) and validation cohorts (n=2054). Cox regression model identified predictors of the primary outcome:
all‐cause mortality, myocardial infarction, or stroke at a mean follow‐up of 2.3 years. The hazard ratio of each variable was rounded to the nearest
integer to construct score weights. A score 0 to 4 defined low‐risk, 5 to 6 intermediate‐risk and ≥7 high‐risk. The following variables were retained in the final model:
age, residence, body mass index, on‐treatment heart rate and BP, prior myocardial infarction, heart failure, stroke/transient ischemic attack, smoking, diabetes, peripheral arterial disease, and chronic kidney disease.
The primary outcome occurred in 2.9\% of the low‐risk group, 6.5\% of the intermediate‐risk group, and 18.0\% of the high‐risk group (P for trend <0.0001).
The model was good at discriminating those who had an event versus those who did not (C‐statistic=0.75). The model performed well in a validation cohort (C‐statistic=0.77).\cr
CONCLUSIONS: Readily available clinical variables can rapidly stratify patients with hypertension and chronic stable CAD into useful risk categories.
}
\references{
Bavry AA, et al. Simple integer risk score to determine prognosis of patients with hypertension and chronic stable coronary artery disease. J Am Heart Assoc. 2013 Aug 15;2(4):e000205. doi: 10.1161/JAHA.113.000205. PMID: 23948642; PMCID: PMC3828777.
}
