% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reach_score.R
\name{reach_score_next_cv}
\alias{reach_score_next_cv}
\alias{reach_score_cv_death}
\alias{reach_score_cv_death_formula}
\alias{reach_score_next_cv_formula}
\title{Calculate Reach-Score_Death / Reach-Score_next_CV}
\usage{
reach_score_next_cv(sex, age, bmi=NA, diabetic=NA, smoker=NA,
vasc=NA, cv_event=NA, chf=NA, af=NA, statin=NA, asa=NA,
region_EE_or_ME = FALSE, region_jap_aust = FALSE)
reach_score_cv_death(sex, age, bmi=NA, diabetic=NA, smoker=NA,
vasc=NA, cv_event=NA, chf=NA, af=NA, statin=NA, asa=NA,
region_EE_or_ME = FALSE, region_jap_aust = FALSE)
reach_score_cv_death_formula(sex, age, bmi=NA, diabetic=NA, smoker=NA,
vasc=NA, cv_event=NA, chf=NA, af=NA, statin=NA, asa=NA,
region_EE_or_ME = FALSE, region_jap_aust = FALSE)
reach_score_next_cv_formula(sex, age, bmi=NA, diabetic=NA, smoker=NA,
vasc=NA, cv_event=NA, chf=NA, af=NA, statin=NA, asa=NA,
region_EE_or_ME = FALSE, region_jap_aust = FALSE)
}
\arguments{
\item{sex}{gender; categorical [female|male]}

\item{age}{age; integer [years]}

\item{bmi}{body mass index; numeric [kg/m^2]}

\item{diabetic}{diabetic status of individual; numeric [1|0]; ("1"=diabetic;"0"=non-diabetic)}

\item{smoker}{information on current self-reported smoking status; numeric [1|0]; ("1"=smoker;"0"=non-smoker)}

\item{vasc}{Number of vascular beds involved in previously diagnosed vascular disease; numeric [0-3]}

\item{cv_event}{information if individual had a cardiovascular event in the past year; numeric [1|0]; ("1"=yes;"0"=no)}

\item{chf}{information if individual has a congestive heart failure (CHF); numeric [1|0]; ("1"=yes;"0"=no)}

\item{af}{information if individual has atrial fibrillation (AF); numeric [1|0]; ("1"=yes;"0"=no)}

\item{statin}{information if individual is on a statin medication; numeric [1|0]; ("1"=yes;"0"=no)}

\item{asa}{information if individual is on a acetylsalicylic acid medication; numeric [1|0]; ("1"=yes;"0"=no)}

\item{region_EE_or_ME}{Geographical region membership in East Europe or Middel East; logical [TRUE|FALSE]}

\item{region_jap_aust}{Geographical region membership in Japan or Australia; logical [TRUE|FALSE]}
}
\value{
A vector of the calculated risk per record.
}
\description{
This function takes necessary parameters to calculate the Reach-Score. You can choose between a calculation based on a score sheet or based on a cox model (formula).
}
\details{
A risk model to predict secondary cardiovascular events and cardiovascular death in outpatients with established atherothrombotic disease.
Tradiotional risk factors, burden of disease, lack of treatment, and geographic location all are related to an increase risk of subsequent
cardiovascular morbidity and cardiovascular mortality. \cr
\cr
Abstract: \cr
Background:\cr
Prediction models for cardiovascular events and cardiovascular death in patients with established cardiovascular disease are not generally available.\cr
Methods: \cr
Participants from the prospective REduction of Atherothrombosis for Continued Health (REACH) Registry provided a global outpatient population with known cardiovascular disease at entry.
Cardiovascular prediction models were estimated from the 2-year follow-up data of 49,689 participants from around the world.\cr
Results:\cr
A developmental prediction model was estimated from 33,419 randomly selected participants (2394 cardiovascular events with 1029 cardiovascular deaths) from the pool of 49,689.
The number of vascular beds with clinical disease, diabetes, smoking, low body mass index, history of atrial fibrillation, cardiac failure,
and history of cardiovascular event(s) <1 year before baseline examination increased risk of a subsequent cardiovascular event.
Statin (hazard ratio 0.75; 95\% confidence interval, 0.69-0.82) and acetylsalicylic acid therapy (hazard ratio 0.90; 95\% confidence interval, 0.83-0.99)
also were significantly associated with reduced risk of cardiovascular events. The prediction model was validated
in the remaining 16,270 REACH subjects (1172 cardiovascular events, 494 cardiovascular deaths). Risk of cardiovascular death was similarly estimated with the same set of risk factors.
Simple algorithms were developed for prediction of overall cardiovascular events and for cardiovascular death.\cr
Conclusions:\cr
This study establishes and validates a risk model to predict secondary cardiovascular events and cardiovascular death in outpatients with established atherothrombotic disease.
Traditional risk factors, burden of disease, lack of treatment, and geographic location all are related to an increased risk of subsequent cardiovascular morbidity and cardiovascular mortality.

\itemize{
\item{Reduction of Atherothrombosis for Continued Health (REACH) Registry}
\item{established in 2003}
\item{Patients from all over the world}
\item{applicable for patients with age >= 45}
\item{Study enrolled between Dec. 2003 nad June 2004}
\item{20-month risk prediction for cardiovascular death and next cardiovascular event}
\item{applicable for Patients with knwon atherothrombotic disease or at high risk to developing atherothrombosis}
}
}
\references{
Wilson. Peter W. F., et al. "An International Model to Predict Recurrent Cardiovascular Disease." The American Journal of Medicine (2012) 125, 695-703.
}
