% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESC_Score2_OP.R
\name{ESC_Score2_OP_table}
\alias{ESC_Score2_OP_table}
\alias{ESC_Score2_OP_formula}
\title{Calculate ESC-Score2 OP table and formula Version}
\usage{
ESC_Score2_OP_table(sex, age, totchol, hdl, sbp, smoker, risk = c("low","moderate","high","very high"), mmol = FALSE)

ESC_Score2_OP_formula(sex, age, totchol, hdl, sbp, smoker, diabetic, risk = c("low","moderate","high","very high"), mmol = FALSE)
}
\arguments{
\item{sex}{a character vector indicating the sex of the person. Values: "female", "male"}

\item{age}{a numeric vector with the age of persons given as years}

\item{totchol}{a numeric vector; Cholesterol values given in mg/dL or mmol/L. If unit is mg/dL set  the argument mmol to FALSE}

\item{hdl}{a numeric vector; HDL Cholesterol values given in mg/dL or mmol/L. If unit is mg/dL set  the argument mmol to FALSE}

\item{sbp}{a numeric vector with the systolic blood pressure of persons given as mmHg}

\item{smoker}{a numeric vector. Smoker = 1, non-smoker = 0. A smoker was defined as current self-reported smoker.}

\item{risk}{logical. Choose if which risk chart is used for calculation}

\item{mmol}{logical. Is Cholesterol given as mmol/L (TRUE) or mg/dL (FALSE).}

\item{diabetic}{a numeric vector. Diabetic status of individual. 1 = diabetic = 1, non-diabetic = 0.}
}
\value{
A vector of calculated risks of persons.
}
\description{
This function takes necessary parameters to calculate the ESC-Score2-OP to estimate 10-year risk of cardiovascular disease in Europe in individuals aged over 70 years in four geographical risk regions.
}
\details{
Aims:
The aim of this study was to derive and validate the SCORE2-Older Persons (SCORE2-OP) risk model to estimate 5- and 10-year risk of cardiovascular disease (CVD)
in individuals aged over 70 years in four geographical risk regions.\cr
Methods and results:\cr
Sex-specific competing risk-adjusted models for estimating CVD risk (CVD mortality, myocardial infarction, or stroke) were derived in individuals
aged over 65 without pre-existing atherosclerotic CVD from the Cohort of Norway (28 503 individuals, 10 089 CVD events).
Models included age, smoking status, diabetes, systolic blood pressure, and total- and high-density lipoprotein cholesterol.
Four geographical risk regions were defined based on country-specific CVD mortality rates. Models were recalibrated to each region using region-specific estimated CVD
incidence rates and risk factor distributions. For external validation, we analysed data from 6 additional study populations {338 615 individuals, 33 219 CVD validation cohorts,
C-indices ranged between 0.63 [95\% confidence interval (CI) 0.61 - 0.65] and 0.67 (0.64 – 0.69)}.
Regional calibration of expected-vs.-observed risks was satisfactory. For given risk factor profiles, there was substantial variation across the four risk regions
in the estimated 10-year CVD event risk.\cr
Conclusions:\cr
The competing risk-adjusted SCORE2-OP model was derived, recalibrated, and externally validated to estimate 5- and 10-year CVD risk in older adults (aged 70 years or older)
in four geographical risk regions. These models can be used for communicating the risk of CVD and potential benefit from risk factor treatment
and may facilitate shared decision-making between clinicians and patients in CVD risk management in older persons.
}
\references{
SCORE2-OP working group and ESC Cardiovascular risk collaboration,
SCORE2-OP risk prediction algorithms: estimating incident cardiovascular event risk in older persons in four geographical risk regions,
European Heart Journal, Volume 42, Issue 25, 1 July 2021, Pages 2455–2467, https://doi.org/10.1093/eurheartj/ehab312
}
