% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reach_score.R
\name{reach_score_next_cv}
\alias{reach_score_next_cv}
\title{Calculate Reach-Score_Death / Reach-Score_next_CV}
\usage{
reach_score_next_cv(sex=NA, age=NA, bmi=NA, diabetic=NA, smoker=NA, vasc=NA, cv_event=NA, chf=NA, af=NA, statin=NA, ass=NA,region_EE_or_ME=NA, region_jap_aust = NA)
reach_score_death(sex=NA, age=NA, bmi=NA, diabetic=NA, smoker=NA, vasc=NA, cv_event=NA, chf=NA, af=NA, statin=NA, ass=NA, region_EE_or_ME = NA, region_jap_aust = NA)
reach_score_cv_death_formula(sex=NA, age=NA, smoker=NA, diabetic=NA, bmi=NA, vasc=NA, cv_event=NA, chf=NA, af=NA, statin=NA, ass=NA, region_EE_or_ME=NA, region_jap_aust = NA)
reach_score_next_cv_formula(sex=NA, age=NA, smoker=NA, diabetic=NA, bmi=NA, vasc=NA, cv_event=NA, chf=NA, af=NA, statin=NA,ass=NA, region_EE_or_ME=NA, region_jap_aust = NA)
}
\arguments{
\item{sex}{a character vector indicating the sex of the person. Values: "female", "male"}

\item{age}{a numeric vector with the age of persons given as years}

\item{bmi}{a numeric vector; Body Mass Index in kg/m^2}

\item{diabetic}{a numeric vector indicating whether a person is diabetic. Values: yes = 1; no = 0.}

\item{smoker}{a numeric vector. Smoker = 1, non-smoker = 0. A smoker was defined as current self-reported smoker.}

\item{vasc}{a numeric vector; Number of vascular beds involved in previously diagnosed vascular disease. Number from 1 to 3}

\item{cv_event}{a numeric vector; cardiovascular event in past year. 1 = yes, 0 = no}

\item{chf}{a numeric vector indicating whether a person had a Congestive heart failure. Values: yes = 1; no = 0.}

\item{statin}{a numeric vector; indicating whether a person is on a statin therapy. 1 = yes, 0 = no}

\item{ass}{a numeric vector; indicating whether a person is on a ASS therapy. 1 = yes, 0 = no.}

\item{region_EE_or_ME}{a logical vector; Geographical region membership in East Europe or Middel East}

\item{region_jap_aust}{a logical vector; Geographical region membership in Japan or Australia}

\item{afa}{numeric vector; Atrial fibrillation. 1 = yes, 0 = no}
}
\value{
A vector of the calculated risk per record.
}
\description{
This function takes necessary parameters to calculate the Reach-Score. You can choose between a calculation based on a score sheet or based on a cox model (formula).
}
\details{
A risk model to predict secondary cardiovascular events and cardiovascular death in outpatients with established atherothrombotic disease.
Tradiotional risk factors, burden of disease, lack of treatment, and geographic location all are related to an increase risk of subsequent cardiovascular morbidity and cardiovascular mortality.
}
\references{
Wilson. Peter W. F., et al. "An International Model to Predict Recurrent Cardiovascular Disease." The American Journal of Medicine (2012) 125, 695-703.
}
