% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reach_Score.R
\name{reach_score_next_cv}
\alias{reach_score_next_cv}
\title{Calculate Reach-Score_Death / Reach-Score_next_CV}
\usage{
reach_score_next_cv(gender=NA, age=NA, bmi=NA, diab=NA, smoker=NA, vasc=NA, cv_event=NA, chf=NA, af=NA,
statin=NA, ass=NA, calc_cv_event = FALSE, khk = NA, mvcad = NA, pavk = NA,  stroke = NA, date_mi = NA,
date_invest = NA, mi = NA, redcap_data = TRUE, data = NULL)

reach_score_death_new(gender=NA, age=NA, bmi=NA, diab=NA, smoker=NA, vasc=NA, cv_event=NA, chf=NA, af=NA,
statin=NA, ass=NA, calc_cv_event = FALSE, khk = NA, mvcad = NA, pavk = NA,  stroke = NA, date_mi = NA,
date_invest = NA, mi = NA, redcap_data = TRUE, data = NULL)
}
\arguments{
\item{gender}{Geschlecht REDCap-ID: varid_549}

\item{age}{Alter REDCap-ID: varid_1891}

\item{bmi}{Body Mass Index (BMI) REDCap-ID: varid_2265}

\item{diab}{Diabetes Mellitus REDCap-ID: varid_558}

\item{smoker}{Raucher REDCap-ID: varid_561}

\item{vasc}{Anzahl der Gefäßbetten mit klinischer Erkrankung (Zerebrovaskuläre Krankheit (TIA, Schlaganfall), koronare Herzkrankheit, pAVK)}

\item{cv_event}{Kardiovaskuläres Event (Myokardinfarkt und zerebrovaskulären Ereignis oder kardiovaskulärer Tod) im letzten Jahr}

\item{chf}{Kategorie Ejektionsfraktion des LV (allgemein)  REDCap-ID: varid_1743}

\item{af}{Vorhofflimmern-/flattern REDCap-ID: varid_576}

\item{statin}{Statin REDCap-ID: varid_693}

\item{ass}{ASS REDCap-ID: varid_695}

\item{calc_cv_event}{logisch; wenn TRUE, werden 'CV_event' und 'vasc' automatisch berechnet. 'khk', 'mvcad', 'pavk', 'stroke', 'date_mi', 'date_invest', 'mi', müssen der Funktion zuätzlich übergeben werden}

\item{khk}{Koronare Herzkrankheit REDCap-ID: varid_569}

\item{mvcad}{Mehrgefäßerkrankung REDCap-ID: varid_1244}

\item{pavk}{PAVK REDCap-ID: varid_610}

\item{stroke}{Schlaganfall/TIA REDCap-ID: varid_613}

\item{date_mi}{Datum Myokardinfarkt REDCap-ID varid_2397}

\item{date_invest}{Datum der Untersuchung/Befragung REDCap-ID: varid_547}

\item{mi}{Z.n. Myokardinfarkt REDCap-ID: varid_570}

\item{redcap_data}{logisch; wenn TRUE benutzt die Funktion den ihr in 'data' übergebenen Dataframe und führt die Berechnung durch. Einzelne Varibalen müssen nicht als Argumente angegeben werden.}

\item{data}{Datensatz mit allen Variablen aus RedCap die für die Berechung benötigt werden. Muss angegben werden wenn 'redcap_data' TRUE.}
}
\value{
A vector of the calculated risk per record.
}
\description{
This function takes necessary parameters to calculate the Reach-Score
}
\details{
Prognosemodell entwickelt zur Sekundärprävention für Patienten mit etablierten kardiovaskulären Erkrankungen, zur Vorhersage von kardiovaskulären Ereignissen oder Tod über einen Zeitraum von 20 Monaten (Wilson et al., 2012).
Der Score wird verwendet, indem jedem Risikoindikator Punkte zugewiesen, die Punkte addiert oder subtrahiert und das prozentuale Risiko für das nächste kardiovaskuläre Ereignis oder den kardiovaskulären Tod bestimmt werden.
Der Score wurde aufgeteilt in zwei einzelne Scores, die zum einen das kardiovaskuläre Risiko und den Tod nach 20 Monaten berechnen. Das prozentuale Risiko unterscheidet sich in beiden Scores.
}
\note{
Die Funktion unterstützt mehrere Varianten für die Berechung der Scores:

\itemize{
\item{Variante 1 (calc_cv_event = FALSE & redcap_data = FALSE):}{

Alle Argumente von 'gender' bis 'ass' müssen der Funktion übergeben werden. 'cv_event' und 'vasc' muss manuell berechnet werden, da keine entsprechenden Felder in REDCap vorhanden sind.
}
\item{Variante 2 (calc_cv_event = TRUE & redcap_data = FALSE):}{

'cv_event' und 'vasc' werden von der Funktion berechnet. Zusätzlich müssen aber die Argumente 'khk', 'mvcad', 'pavk', 'stroke', 'date_mi', 'date_invest', 'mi' der Funktion übergeben werden.
}
\item{Variante 3 (calc_cv_event = FALSE & redcap_data = TRUE):}{

Wenn 'redcap_data' TRUE und ein Dataframe mit allen benötigten Variablen aus REDCAp (alle in den Argumenten beschriebenen varids) übergeben wird, berechnet die Funktion die Scores ohne das weitere Argumente angegeben werden müssen.
}
}
}
