% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reach_score.R
\name{reach_score_next_cv}
\alias{reach_score_next_cv}
\alias{reach_score_cv_death}
\alias{reach_score_cv_death_formula}
\alias{reach_score_next_cv_formula}
\title{Calculate Reach-Score_Death / Reach-Score_next_CV}
\usage{
reach_score_next_cv(sex, age, bmi=NA, diabetic=NA, smoker=NA,
vasc=NA, cv_event=NA, chf=NA, af=NA, statin=NA, asa=NA,
region_EE_or_ME = FALSE, region_jap_aust = FALSE)
reach_score_cv_death(sex, age, bmi=NA, diabetic=NA, smoker=NA,
vasc=NA, cv_event=NA, chf=NA, af=NA, statin=NA, asa=NA,
region_EE_or_ME = FALSE, region_jap_aust = FALSE)
reach_score_cv_death_formula(sex, age, bmi=NA, diabetic=NA, smoker=NA,
vasc=NA, cv_event=NA, chf=NA, af=NA, statin=NA, asa=NA,
region_EE_or_ME = FALSE, region_jap_aust = FALSE)
reach_score_next_cv_formula(sex, age, bmi=NA, diabetic=NA, smoker=NA,
vasc=NA, cv_event=NA, chf=NA, af=NA, statin=NA, asa=NA,
region_EE_or_ME = FALSE, region_jap_aust = FALSE)
}
\arguments{
\item{sex}{a character vector indicating the sex of the person. Values: "female", "male"}

\item{age}{a numeric vector with the age of persons given as years}

\item{bmi}{a numeric vector; Body Mass Index in kg/m^2}

\item{diabetic}{a numeric vector indicating whether a person is diabetic. Values: yes = 1; no = 0.}

\item{smoker}{a numeric vector. A smoker was defined as >= 5 cigarettes per day on average within the last month. Smoker = 1, non-smoker = 0.}

\item{vasc}{a numeric vector; Number of vascular beds involved in previously diagnosed vascular disease. Number from 1 to 3}

\item{cv_event}{a numeric vector; cardiovascular event in past year. 1 = yes, 0 = no}

\item{chf}{a numeric vector indicating whether a person had a Congestive heart failure. Values: yes = 1; no = 0.}

\item{af}{numeric vector; Atrial fibrillation. 1 = yes, 0 = no}

\item{statin}{a numeric vector; indicating whether a person is on a statin therapy. 1 = yes, 0 = no}

\item{asa}{information if individual is on a acetylsalicylic acid medication; numeric ("1"=yes;"0"=no)}

\item{region_EE_or_ME}{a logical vector; Geographical region membership in East Europe or Middel East}

\item{region_jap_aust}{a logical vector; Geographical region membership in Japan or Australia}
}
\value{
A vector of the calculated risk per record.
}
\description{
This function takes necessary parameters to calculate the Reach-Score. You can choose between a calculation based on a score sheet or based on a cox model (formula).
}
\details{
A risk model to predict secondary cardiovascular events and cardiovascular death in outpatients with established atherothrombotic disease.
Tradiotional risk factors, burden of disease, lack of treatment, and geographic location all are related to an increase risk of subsequent
cardiovascular morbidity and cardiovascular mortality.
\itemize{
\item{Reduction of Atherothrombosis for Continued Health (REACH) Registry}
\item{established in 2003}
\item{Patients from all over the world}
\item{applicable for patients with age >= 45}
\item{Study enrolled between Dec. 2003 nad June 2004}
\item{20-month risk prediction for cardiovascular death and next cardiovascular event}
\item{applicable for Patients with knwon atherothrombotic disease or at high risk to developing atherothrombosis}
}
}
\references{
Wilson. Peter W. F., et al. "An International Model to Predict Recurrent Cardiovascular Disease." The American Journal of Medicine (2012) 125, 695-703.
}
