% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRA2P_Score.R
\name{TRA2P_Score}
\alias{TRA2P_Score}
\title{Calculate TRA2P-Score}
\usage{
TRA2P_Score_new_2(CHF = NA, AH = NA, Alter = NA, Diabetes = NA, Schlaganfall = NA, Bypass_OP = NA, andere_OP = NA, eGFR = NA, Raucher = NA, pavk = NA, redcap_data = FALSE, data = NULL)
}
\arguments{
\item{CHF}{Kategorie Ejektionsfraktion des LV (allgemein) REDCap-ID: varid_1743}

\item{AH}{Arterielle Hypertonie REDCap-ID: varid_559}

\item{Alter}{Alter REDCap-ID: varid_1891}

\item{Diabetes}{Diabetes Mellitus REDCap-ID: varid_558}

\item{Schlaganfall}{Schlaganfall/TIA REDCap-ID: varid_613}

\item{Bypass_OP}{Koronare Bypass-OP REDCap-ID: varid_584}

\item{andere_OP}{Sonstige Gefäss-Operation REDCap-ID: varid_586}

\item{eGFR}{eGFR MDRD REDCap-ID: varid_2629}

\item{Raucher}{Raucher REDCap-ID: varid_561}

\item{pavk}{PAVK REDCap-ID: varid_610}

\item{redcap_data}{logisch; wenn TRUE benutzt die Funktion den ihr in 'data' übergebenen Dataframe und führt die Berechnung durch. Einzelne Varibalen müssen nicht als Argumente angegeben werden.}

\item{data}{Datensatz mit allen Variablen aus RedCap die für die Berechung benötigt werden. Muss angegben werden wenn 'redcap_data' TRUE.}
}
\value{
A vector of the calculated risk per record.
}
\description{
This function takes necessary parameters to calculate the TRA2P-Score
}
\details{
Der Score sagt ein drei Jahres Risiko für MI, kardiovaskulären Tod und ischämischen Schlaganfall voraus (Bohula et al., 2016).
}
